backend/verify/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
dlib==19.24.2
